python/restapi_project/requirements.txt
sqlalchemy
twisted
pyopenssl
PyYaml
nomad
